+++ rpc/core/src/model/message.rs
@@ -732,6 +732,69 @@ impl Deserializer for SubmitTransactionResponse {
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
+pub struct SubmitTransactionReplacementRequest {
+    pub transaction: RpcTransaction,
+}
+
+impl SubmitTransactionReplacementRequest {
+    pub fn new(transaction: RpcTransaction) -> Self {
+        Self { transaction }
+    }
+}
+
+impl Serializer for SubmitTransactionReplacementRequest {
+    fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
+        store!(u16, &1, writer)?;
+        serialize!(RpcTransaction, &self.transaction, writer)?;
+
+        Ok(())
+    }
+}
+
+impl Deserializer for SubmitTransactionReplacementRequest {
+    fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
+        let _version = load!(u16, reader)?;
+        let transaction = deserialize!(RpcTransaction, reader)?;
+
+        Ok(Self { transaction })
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct SubmitTransactionReplacementResponse {
+    pub transaction_id: RpcTransactionId,
+    pub replaced_transaction: RpcTransaction,
+}
+
+impl SubmitTransactionReplacementResponse {
+    pub fn new(transaction_id: RpcTransactionId, replaced_transaction: RpcTransaction) -> Self {
+        Self { transaction_id, replaced_transaction }
+    }
+}
+
+impl Serializer for SubmitTransactionReplacementResponse {
+    fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
+        store!(u16, &1, writer)?;
+        store!(RpcTransactionId, &self.transaction_id, writer)?;
+        serialize!(RpcTransaction, &self.replaced_transaction, writer)?;
+
+        Ok(())
+    }
+}
+
+impl Deserializer for SubmitTransactionReplacementResponse {
+    fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
+        let _version = load!(u16, reader)?;
+        let transaction_id = load!(RpcTransactionId, reader)?;
+        let replaced_transaction = deserialize!(RpcTransaction, reader)?;
+
+        Ok(Self { transaction_id, replaced_transaction })
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize)]
+#[serde(rename_all = "camelCase")]
 pub struct GetSubnetworkRequest {
     pub subnetwork_id: RpcSubnetworkId,
 }
@@ -1766,20 +1829,25 @@ impl Deserializer for GetConnectionsRequest {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GetConnectionsResponse {
-    pub active_connections: u32,
+    pub clients: u32,
+    pub peers: u16,
 }
 
 impl Serializer for GetConnectionsResponse {
     fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        store!(u32, &self.active_connections, writer)?;
+        store!(u16, &1, writer)?;
+        store!(u32, &self.clients, writer)?;
+        store!(u16, &self.peers, writer)?;
         Ok(())
     }
 }
 
 impl Deserializer for GetConnectionsResponse {
     fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let active_connections = load!(u32, reader)?;
-        Ok(Self { active_connections })
+        let _version = load!(u16, reader)?;
+        let clients = load!(u32, reader)?;
+        let peers = load!(u16, reader)?;
+        Ok(Self { clients, peers })
     }
 }