+++ rpc/wrpc/server/src/router.rs
@@ -69,6 +69,7 @@ impl Router {
                 Shutdown,
                 SubmitBlock,
                 SubmitTransaction,
+                SubmitTransactionReplacement,
                 Unban,
             ]
         );