+++ wallet/core/src/utxo/settings.rs
@@ -4,128 +4,104 @@
 //!
 
 use crate::imports::*;
+use kaspa_consensus_core::mass::Kip9Version;
 
 #[derive(Debug)]
-pub struct Inner {
+pub struct NetworkParams {
     pub coinbase_transaction_maturity_period_daa: AtomicU64,
     pub coinbase_transaction_stasis_period_daa: u64,
     pub user_transaction_maturity_period_daa: AtomicU64,
-    pub mass_combination_strategy: MassCombinationStrategy,
+    pub kip9_version: Kip9Version,
     pub additional_compound_transaction_mass: u64,
-}
-
-#[derive(Debug, Clone)]
-pub struct NetworkParams {
-    inner: Arc<Inner>,
 }
 
 impl NetworkParams {
     #[inline]
     pub fn coinbase_transaction_maturity_period_daa(&self) -> u64 {
-        self.inner.coinbase_transaction_maturity_period_daa.load(Ordering::Relaxed)
+        self.coinbase_transaction_maturity_period_daa.load(Ordering::Relaxed)
     }
 
     #[inline]
     pub fn coinbase_transaction_stasis_period_daa(&self) -> u64 {
-        self.inner.coinbase_transaction_stasis_period_daa
+        self.coinbase_transaction_stasis_period_daa
     }
 
     #[inline]
     pub fn user_transaction_maturity_period_daa(&self) -> u64 {
-        self.inner.user_transaction_maturity_period_daa.load(Ordering::Relaxed)
+        self.user_transaction_maturity_period_daa.load(Ordering::Relaxed)
     }
 
     #[inline]
-    pub fn mass_combination_strategy(&self) -> MassCombinationStrategy {
-        self.inner.mass_combination_strategy
+    pub fn kip9_version(&self) -> Kip9Version {
+        self.kip9_version
     }
 
     #[inline]
     pub fn additional_compound_transaction_mass(&self) -> u64 {
-        self.inner.additional_compound_transaction_mass
+        self.additional_compound_transaction_mass
     }
 
     pub fn set_coinbase_transaction_maturity_period_daa(&self, value: u64) {
-        self.inner.coinbase_transaction_maturity_period_daa.store(value, Ordering::Relaxed);
+        self.coinbase_transaction_maturity_period_daa.store(value, Ordering::Relaxed);
     }
 
     pub fn set_user_transaction_maturity_period_daa(&self, value: u64) {
-        self.inner.user_transaction_maturity_period_daa.store(value, Ordering::Relaxed);
+        self.user_transaction_maturity_period_daa.store(value, Ordering::Relaxed);
     }
 }
 
-lazy_static::lazy_static! {
-    pub static ref MAINNET_NETWORK_PARAMS: NetworkParams = NetworkParams {
-        inner: Arc::new(Inner {
-            coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
-            coinbase_transaction_stasis_period_daa: 50,
-            user_transaction_maturity_period_daa: AtomicU64::new(10),
-            mass_combination_strategy: MassCombinationStrategy::Max,
-            additional_compound_transaction_mass: 0,
-        }),
-    };
-}
+static MAINNET_NETWORK_PARAMS: LazyLock<NetworkParams> = LazyLock::new(|| NetworkParams {
+    coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
+    coinbase_transaction_stasis_period_daa: 50,
+    user_transaction_maturity_period_daa: AtomicU64::new(10),
+    kip9_version: Kip9Version::Beta,
+    additional_compound_transaction_mass: 100,
+});
 
-lazy_static::lazy_static! {
-    pub static ref TESTNET10_NETWORK_PARAMS: NetworkParams = NetworkParams {
-        inner: Arc::new(Inner {
-            coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
-            coinbase_transaction_stasis_period_daa: 50,
-            user_transaction_maturity_period_daa: AtomicU64::new(10),
-            mass_combination_strategy: MassCombinationStrategy::Max,
-            additional_compound_transaction_mass: 0,
-        }),
-    };
-}
+static TESTNET10_NETWORK_PARAMS: LazyLock<NetworkParams> = LazyLock::new(|| NetworkParams {
+    coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
+    coinbase_transaction_stasis_period_daa: 50,
+    user_transaction_maturity_period_daa: AtomicU64::new(10),
+    kip9_version: Kip9Version::Beta,
+    additional_compound_transaction_mass: 100,
+});
 
-lazy_static::lazy_static! {
-    pub static ref TESTNET11_NETWORK_PARAMS: NetworkParams = NetworkParams {
-        inner: Arc::new(Inner {
-            coinbase_transaction_maturity_period_daa: AtomicU64::new(1_000),
-            coinbase_transaction_stasis_period_daa: 500,
-            user_transaction_maturity_period_daa: AtomicU64::new(100),
-            mass_combination_strategy: MassCombinationStrategy::Max,
-            additional_compound_transaction_mass: 100,
-        }),
-    };
-}
+static TESTNET11_NETWORK_PARAMS: LazyLock<NetworkParams> = LazyLock::new(|| NetworkParams {
+    coinbase_transaction_maturity_period_daa: AtomicU64::new(1_000),
+    coinbase_transaction_stasis_period_daa: 500,
+    user_transaction_maturity_period_daa: AtomicU64::new(100),
+    kip9_version: Kip9Version::Alpha,
+    additional_compound_transaction_mass: 100,
+});
 
-lazy_static::lazy_static! {
-    pub static ref SIMNET_NETWORK_PARAMS: NetworkParams = NetworkParams {
-        inner: Arc::new(Inner {
-            coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
-            coinbase_transaction_stasis_period_daa: 50,
-            user_transaction_maturity_period_daa: AtomicU64::new(10),
-            mass_combination_strategy: MassCombinationStrategy::Max,
-            additional_compound_transaction_mass: 0,
-        }),
-    };
-}
+static SIMNET_NETWORK_PARAMS: LazyLock<NetworkParams> = LazyLock::new(|| NetworkParams {
+    coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
+    coinbase_transaction_stasis_period_daa: 50,
+    user_transaction_maturity_period_daa: AtomicU64::new(10),
+    kip9_version: Kip9Version::Alpha,
+    additional_compound_transaction_mass: 0,
+});
 
-lazy_static::lazy_static! {
-    pub static ref DEVNET_NETWORK_PARAMS: NetworkParams = NetworkParams {
-        inner: Arc::new(Inner {
-            coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
-            coinbase_transaction_stasis_period_daa: 50,
-            user_transaction_maturity_period_daa: AtomicU64::new(10),
-            mass_combination_strategy: MassCombinationStrategy::Max,
-            additional_compound_transaction_mass: 0,
-        }),
-    };
-}
+static DEVNET_NETWORK_PARAMS: LazyLock<NetworkParams> = LazyLock::new(|| NetworkParams {
+    coinbase_transaction_maturity_period_daa: AtomicU64::new(100),
+    coinbase_transaction_stasis_period_daa: 50,
+    user_transaction_maturity_period_daa: AtomicU64::new(10),
+    kip9_version: Kip9Version::Beta,
+    additional_compound_transaction_mass: 0,
+});
 
-impl From<NetworkId> for NetworkParams {
-    fn from(value: NetworkId) -> Self {
+impl NetworkParams {
+    pub fn from(value: NetworkId) -> &'static NetworkParams {
         match value.network_type {
-            NetworkType::Mainnet => MAINNET_NETWORK_PARAMS.clone(),
+            NetworkType::Mainnet => &MAINNET_NETWORK_PARAMS,
             NetworkType::Testnet => match value.suffix {
-                Some(10) => TESTNET10_NETWORK_PARAMS.clone(),
-                Some(11) => TESTNET11_NETWORK_PARAMS.clone(),
+                Some(10) => &TESTNET10_NETWORK_PARAMS,
+                Some(11) => &TESTNET11_NETWORK_PARAMS,
                 Some(x) => panic!("Testnet suffix {} is not supported", x),
                 None => panic!("Testnet suffix not provided"),
             },
-            NetworkType::Devnet => DEVNET_NETWORK_PARAMS.clone(),
-            NetworkType::Simnet => SIMNET_NETWORK_PARAMS.clone(),
+            NetworkType::Devnet => &DEVNET_NETWORK_PARAMS,
+            NetworkType::Simnet => &SIMNET_NETWORK_PARAMS,
         }
     }
 }